Check flag first and simplify
